Added navigation and a folder for pages
